pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "rfab-harness"
version = "0.1.0"
description = "Harness for RFAntibody antibody design campaigns"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "biopython>=1.80",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
rfab = "harness.cli:main"

[tool.setuptools.packages.find]
include = ["harness*"]
